refactor: rewrite the register template
This is more in line with what the Figma design envisioned.
I prepared the template for translations. You can update your `.env`
file for German strings going forward.

https://laravel.com/docs/11.x/localization#configuring-the-locale

It's unfortunate that we are using TailwindCSS as this carries with it
rulesets using `!important` which cannot be easily overwritten further.
I'd know how to do that in (S)CSS but not in TailwindCSS.

So instead I found creative workarounds.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>