feat: resolve conflicts between twbs and tw
Both used a class called .collapse. When I add .visible I beat
TailwindCSS to behave.

With that out of the way I went ahead and ensured that the icons always
have an accessible name. Therefore I hide the icons from Assistive
Technology. For smaller viewports we want to show a label next to the
icon. I hide them on larger ones.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>