feat: populate Organizer of Event
Assuming that the user who can create an Event is an Organizer write the
ID into the column.

For demonstration purposes the Organizer table is seeded with an User.
This is not taking care of setting the permissions, but we don't make
use of permissions at this point in time anyway.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>